Refactor OSTreeRepo to take advantage of object types.
