+++ src/sota_tools/ostree_dir_repo.cc
@@ -1,16 +1,11 @@
 #include "ostree_dir_repo.h"
 
-#include <iomanip>
-#include <iostream>
-#include <sstream>
 #include <string>
 
-#include <boost/filesystem.hpp>
 #include <boost/property_tree/ini_parser.hpp>
 
 #include "logging/logging.h"
 
-using std::string;
 namespace fs = boost::filesystem;
 namespace pt = boost::property_tree;
 
@@ -22,7 +17,7 @@ bool OSTreeDirRepo::LooksValid() const {
     pt::ptree config;
     try {
       pt::read_ini(config_file.string(), config);
-      if (config.get<string>("core.mode") != "archive-z2") {
+      if (config.get<std::string>("core.mode") != "archive-z2") {
         LOG_WARNING << "OSTree repo is not in archive-z2 format";
         return false;
       }
@@ -42,30 +37,8 @@ bool OSTreeDirRepo::LooksValid() const {
 
 OSTreeRef OSTreeDirRepo::GetRef(const std::string &refname) const { return OSTreeRef(*this, refname); }
 
-OSTreeObject::ptr OSTreeDirRepo::GetObject(const uint8_t sha256[32], const OstreeObjectType type) const {
-  return GetObject(OSTreeHash(sha256), type);
-}
-
-OSTreeObject::ptr OSTreeDirRepo::GetObject(const OSTreeHash hash, const OstreeObjectType type) const {
-  otable::const_iterator it;
-  it = ObjectTable.find(hash);
-  if (it != ObjectTable.end()) {
-    return it->second;
-  }
-
-  // TODO: fix?
-  (void)type;
-  std::string exts[] = {".filez", ".dirtree", ".dirmeta", ".commit"};
-  boost::filesystem::path objpath = hash.string().insert(2, 1, '/');
-
-  for (const std::string &ext : exts) {
-    if (fs::is_regular_file((root_ / "/objects/" / objpath).string() + ext)) {
-      OSTreeObject::ptr obj(new OSTreeObject(*this, objpath.string() + ext));
-      ObjectTable[hash] = obj;
-      return obj;
-    }
-  }
-  throw OSTreeObjectMissing(hash);
+bool OSTreeDirRepo::FetchObject(const boost::filesystem::path &path) const {
+  return fs::is_regular_file((root_ / path).string());
 }
 
 // vim: set tabstop=2 shiftwidth=2 expandtab:
