+++ src/sota_tools/ostree_http_repo.h
@@ -1,13 +1,8 @@
 #ifndef SOTA_CLIENT_TOOLS_OSTREE_HTTP_REPO_H_
 #define SOTA_CLIENT_TOOLS_OSTREE_HTTP_REPO_H_
 
-#include <list>
-#include <map>
-
 #include <boost/filesystem.hpp>
 
-#include "ostree_hash.h"
-#include "ostree_object.h"
 #include "ostree_ref.h"
 #include "ostree_repo.h"
 #include "treehub_server.h"
@@ -15,21 +10,16 @@
 class OSTreeHttpRepo : public OSTreeRepo {
  public:
   explicit OSTreeHttpRepo(const TreehubServer* server) : server_(server) {}
+  ~OSTreeHttpRepo() override = default;
 
   bool LooksValid() const override;
-  OSTreeObject::ptr GetObject(OSTreeHash hash, OstreeObjectType type) const override;
-  OSTreeObject::ptr GetObject(const uint8_t sha256[32], OstreeObjectType type) const override;
   OSTreeRef GetRef(const std::string& refname) const override;
-
   const boost::filesystem::path root() const override { return root_.Path(); }
 
  private:
-  bool Get(const boost::filesystem::path& path) const;
-  bool CheckForObject(const OSTreeHash& hash, const std::string& path, OSTreeObject::ptr& object) const;
+  bool FetchObject(const boost::filesystem::path& path) const override;
   static size_t curl_handle_write(void* buffer, size_t size, size_t nmemb, void* userp);
 
-  typedef std::map<OSTreeHash, OSTreeObject::ptr> otable;
-  mutable otable ObjectTable;  // Makes sure that the same commit object is not added twice
   const TreehubServer* server_;
   const TemporaryDirectory root_;
 };
